+++ src/utils/mod.rs
@@ -1,1 +1,3 @@
+#[cfg(test)]
+pub mod xml_eq;
 pub mod yaserde;
